container-app/pyproject.toml
[project]
name = "mcp-finance-server"
version = "1.0.0"
description = "MCP Finance Data Server for company earnings data"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "starlette>=0.40.0",
]

[project.scripts]
mcp-finance-server = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
